Add SmolVLA kernel Programs with golden data and MLIR
List of kernels added with Golden data and MLIR implementation:
- silu (pre-existing template, unchanged)
- elementwise_add, elementwise_mul, elementwise_sub, elementwise_div
- matmul (single-tile 32x32) and matmul_k_chain (K=64, two-tile accumulate)
- rms_norm, softmax, reduction_sum, requant
- rope_frequency, gelu_tanh
- attention (single-tile SDPA) and fused_attention (flash attention, KV=2)

Each Program is self-contained (MLIR string + PyTorch reference +
seeded golden tensor + inline ISA) and validated two ways via
test_programs.py: simulator output vs golden, and IREE compile+run of
the MLIR vs PyTorch reference at import time.

Changes outside npu_model/configs/programs/:
- npu_model/hardware/arch_state.py: byte-level MRF fp8 write (view as
  uint8 to avoid value-cast), required for requant correctness.
- npu_model/configs/isa_definition.py: vpack.bf16.fp8 writes the
  quantized tile as uint8 bytes, pairing with the arch_state fp8 fix.
- scripts/test_programs.py: per-program kernel_tolerance override so
  fp8-quantized kernels (e.g. attention) can opt into a looser tol.

+++ npu_model/hardware/arch_state.py
@@ -131,7 +131,10 @@ class ArchState:
             value.numel()
             == self.cfg.mrf_depth * self.cfg.mrf_width // torch.float8_e4m3fn.itemsize
         )
-        self.mrf[vd].view(torch.float8_e4m3fn)[:] = value.flatten()
+        # Byte-level copy: view target as uint8 and assign, otherwise PyTorch
+        # value-casts uint8→fp8 (e.g. 173 → 176.0, bytes 115) instead of
+        # reinterpreting the bits.
+        self.mrf[vd].view(torch.uint8)[:] = value.flatten()
 
     def read_mrf_fp8(self, vs: int) -> torch.Tensor:
         return (
